whisperme-python/pyproject.toml
[project]
name = "whisperme-backend"
version = "0.1.0"
description = "WhisperMe transcription backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "httpx>=0.25.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "supabase>=2.16.0",
    "pytz>=2023.3",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
] 
